KOSMOS/requirements.txt
PyQt6==6.7.*
PyQt6-WebEngine==6.7.*
SQLAlchemy==2.0.*
feedparser==6.0.*
requests==2.31.*
beautifulsoup4==4.12.*
lxml>=5
newspaper4k==0.9.*
praw==7.7.*
deep-translator==1.11.*
weasyprint==62.*
Pillow>=10
matplotlib==3.8.*
trafilatura
readability-lxml
html2text
langdetect
rapidfuzz
